Rec: introduce PATCH to fix playbacks
This commit introduces a PATCH operation on playback resources. The
client can send the same resource used for creating playbacks in order
to alter the details of a playback.
All fields that are null (or lists that are empty) are ignored.

In case of success the updated resource is returned.