Cli: fix time-keeping in PlayerStateMachine
There was a discrepancy in the time measured by the PlayerStateMachine
as opposed to the time at which events actually occur due to the fact
that time was measured before and after the callback to the client code.

In the callback we often interacted with the back-end service (posting
playbacks) which lead to lower times measured by the PlayerStateMachine.

This patch fixes the problem.

Fixes #12