Rec: Introduce AbstractDetachedDto to replace AbstractDto
Introduce AbstractDetachedDto as base class for API DTOs. In contrast to
AbstractDto, this base class does not retain a reference to the
represented entity. We want to establish a clean split between API layer
and internal representations.
The AbstractDto is marked deprecated and is to be replaced by
AbstractDetachedDto step by step.