Do not expose FQDN of entities in error cases
The NotFoundException includes in the message, which type of entity
could not be found. We use a reflective class object for this and used
to extract the qualifiedName. Now, this is changed to simpleName to not
expose to many details of the implementation.
Exposing the FQDN is less of a security problem and more a matter of
cluttering error messages with unhelpful information.