Add endpoint to POST a Playback from track details
This patch introduces the POST endpoint to add a new Playback. Every
user can only create Playbacks for themselves, obviously. The originally
submitted properties are all stored in a separate lazy-fetched entity.
In the future they will only be relevant for "broken playbacks", i.e.
for playbacks whose track could not be identified.

One can optionally specify a playback timestamp, which will be used for
importing playbacks from a data source or for "offline mode" of any cfm
clients.

Currently the API just presents the data that the user just submitted.
This will be changed in the future and the properties found in the mb
database will be returned.