mbs: Support full parameter for identify call
Clients can get a different response for the identify call when
providing full=true as a parameter. In this case, mbs will not return a
shallow object with references, but rather a deeply nested structure
containing all the details, i.e. the entire recording and release with
their complete artists.

The full parameter is optional and, if left out, defaults to the
previous behaviour.