Identify recordings by title, release, artists
This patch introduces lookup of musicbrainz details for a given
playback. The user can provide artists, recording title and release
title of a playback. The recorder service will forward the information
to mbs and, if found, receive the details of the release group and
recording and their artists respectively.

Internally, these resources are translated into the entity types Artist,
ReleaseGroup and Recording respectively. The service will then check if
they are already in our database or not. If they are, we check if an
update is required. If they are not, they will stored initially.

The "store or update" logic is located in the respective boundaries. The
identification (i.e. consuming mbs and providing the results internally)
is done in the PlaybackIdentifier service.

Furthermore, this patch introduces two new exception types.

InvalidResourceException: should be used in an "assertion-style". That
means we usually do not expect that this Exception is thrown, except if
there is a fatal unexpected error. The Exception can be used to express
that a resource (any object) is invalid due to a field that does not
have the expected state. An example would be a null value in a required
field.

MbsLookupFailedException: simply used to communicate, that the
downstream service (mbs) could not be queried successfully.