Fix issue with initial system user creation
SystemInit tries to create an initial system user, if it does not exist,
or promote an existing user to system user. Therefore, it used to call
the createUser() or promoteToSystemUser() functions of UserBoundary
respectively. Since these methods are now protected and can only be
executed with a user context, this failed and the application could not
initialize a working state from an empty database.

Since Spring only checks authorization for the *first* call into a
service, we simply redirect the calls to createUser() and
promoteToSystemUser() by introducing the methods createUserNoAuthCheck()
and promoteToSystemUserNoAuthCheck(). They just call the originals and
do nothing else. This effectively disables the authorization checks.