Introduce generic 'ListRes'ource and translation
This patch introduces a generic 'ListRes'ource that can be used to
provide any kind of resources bundled in lists on the API.

Previously, we introduced the concept of translation functions which,
for example, take an entity or data model (projection, ...) and
translates them into a "resource" which is a data structure that will be
sent over the wire. This translation includes the following advantages:
* split internal model and external representation, if one changes, the
  other does not have to change
* security, by hiding internal details
* aggregate raw data to provide calculated properties
...

This concept only applied for single resources (potentially nested
though). The ListRes represents a list of elements of the same type.
When we get a list or page of internal objects, we want to translate all
the internal objects and the wrapping list or page into the external
representation.

The ListRes<E, R> incorporates the *TYPES* of the internal and external
representation. But it only holds the *DATA* of the external
representation. The internal type is only necessary to be able to
translate all the inner objects into the external format.

The extension fun <E, R> Page<E> toRes() takes in a function to translate
all the contained Es into Rs and returns a new ListRes wrapping these
Rs.