Store lastUpdated timestamp as long instead of Date
Until now we store the lastUpdated timestamp as Date with
@Temporal(TIMESTAMP) annotation. This had the advantage, that Jackson
can deserialise UNIX timestamps directly into Date objects. However,
this lead to an issue:

Hibernate does *not* instantiate a Date object, but rather a
java.sql.Timestamp, which inherits from Date. But on the resource, we
receive a Date, which makes it virtually impossible to compare the two.
Calling equals() will always fail. Calling before() would only work in
one direction (from the super class to the subtype), which would also be
very confusing.

Trying to make the entity field explicitly a Timestamp fails, because
Hibernate is an asshole. Also I'm too lazy to cover all potential corner
cases when Hibernate decides to be a fucking douche bag and handle
things different internally.

Consequently, we simply store the timestamp as a long and compare
numerically. If needed, we can always convert the long into a date/time
representation.