Fix "artist losing" bug
We used to return "null" if an artist was already stored and did not
need update. The intention was simple: we did not want to bother
Hibernate with unchanged entities, because all returned non-null values
would be persisted afterwards.

But the returned values are ALSO passed along to the caller of the
boundary method. Consequently, all existing and unupdated artists would
be missing in the return value. This caused ReleaseGroups or Recordings
to lack the relations to their artists.

We now always return something, either a new yet unpersisted artist or
the persisted one and rely on Hibernate to figure out which it has to
actually store (which it does correctly, so it seems. Unbelievable, I
know).