Add split implementations of http client
Currently, there is one big client module which is growing with each
added feature. This moo class also holds the base URL of the endpoint.
In this commit, we split the single client module into three modules. A
base module providing generic request subs for get, post, patch and so
on. Two derived modules provide subs for specific requests against
either recorder or mbs.

By this we achieve:
1. better overview in either of the specific clients
2. hopefully faster loading times for the modules, because they are
   smaller
3. we can have different endpoints for recorder and mbs, which might be
   very useful in the future. Not everybody has to run mbs then, for
   example.