Introduce autowire, singletons and beans
This patch introduces two concepts that facilitate writing packages.
Based on Moo, one can now use the subs "autowire" and "singleton" to get
a new instance or always one and the same instance respectively. Any Moo
package can be autowired. In order to get a singleton of a Moo package,
it must include the role "Cfm::Singleton".

Just like typed dependency injection, one can define "beans" which are
just named singletons of anything. The injected thing does not need to
be a Moo class, it can be any value.