Provide roles for Resource and ListResource
The role Res can be included into a class whenever it represents a
"Resource", which in our terms is anything that exists in the
server-side model as well. The role brings along the logic to
deserialise a HASH (originall JSON) representation.

The ListRes is a more specialised role that does the same for list
resources. Since it cannot possibly know what is its content, we pass
along a deserialising sub.