Provide Config as a singleton
Previously, the config was loaded proactively from the CLI module and
evaluated depending for the specific command. This commit adds a
dedicated class for the configuration. With the previously introduced
singleton role, the config can be injected into any moo class.

During the first initialisation, the config is loaded from a config
file. Due to lazy loading, the config will only be loaded if we access
it.