CfmClient pulls config from config singleton
Instead of setting the configuration from the outside, CfmClient now
injects the config singleton and reads the respective configuration
parameters. For this, we pulled the authorization header setting from
the base class to the more special implementation, since different
clients are designed to call different endpoints with potentially
different credentials.