Config holds a hash rather than a complex Config object
Our config is read-only with respects to the config file. Once loaded we
will not modify the options. Thus, we do not have to interact with the
Config::Simple object but can rather just take a hash of configuration
options. That makes it easier to deal with overriding options, as we can
simply replace the value in the hash instead of shadowing it.