Add list command to new modular Cli implementation
The new modular Cli class previously gained the basic command matching
logic. This patch adds the first command "list" to it. It also
introduces the new version of formatters.

There is now a role Formatter in which we hook into the dependency
injection mechanism. Instead of returning a singleton instance of the
Formatter itself (which btw cannot exist, because you cannot instantiate
a role), it figures out which is the wanted implementation and returns
the singleton instance of this subtype. This also implies, that the
Formatter can only be injected with "singleton" but not with "autowire".
You can, however, autowire the subtypes. Furthermore, you can also
inject the subtypes via "singleton" and you will receive the same
instance.