Add serialisation of Resources
This patch enables classes with the role Res (Resource) to serialise
themselves to a hash. The result of the to_hash() sub can be passed to
encode_json(). The implemented mechanism allows classes to define a hash
"projection" which may map field names to a sub that produce a
serialisable representation of the content.

So, if a class is for example pointing to another blessed class, the
projection may simply call to_hash() on that class as well. The result
of to_hash() is always an unblessed hash. It is also possible to
transform values entirely using the projection, for example translating
ints into JSON::true or JSON::false.

The "projection" hash can always be left undefined. In this case, the
fields will be used unaltered in the resulting hash.