Add key-value store to Config
We introduce a conceptual key-value store to the Config. The user can
fill the kv-store with the -s (or --set) flag, like so: -s 'key:value'.
Both key and value are trimmed by the regex. After the flags are pushed
to the Config object, you can ask for the kv-store using the sub
kv_store() which returns a hash with the mapping.

Since this hash is calculated "on demand", you should call the
kv_store() sub only once and hold the result in a variable as long as
you need it.