Provide new simple recording endpoint
This patch removes the old endpoint to retrieve a single Recording with
references to the recording artists. The newly introduced endpoint
provides a view-based resource instead. We can extend the resource if
more things need be shown on the client.

This patch is part of ongoing work to free the recorder of any
persistent musicbrainz entities. The view-based resource may be cached
in the future, but it will not be persisted.