Refactor and update the identify playback endpoint
Firstly, make the PlaybackBoundary a kotlin service and throw away all
the unnecessary weight. It now returns a kotlin Pair of Release and
Recording.
The API is also implemented in kotlin and simplified, so we can throw
away all the old resources and factory. The IdentifiedPlaybackRes is the
new replacement. Note, that this resource is actually different. So this
patch is not pure refactoring but rather updates the endpoint to our new
needs.

In the future, this can be seen as a mere technical endpoint to perform
the identification of a played song. The details must be retrieved in a
separate call. This slight inconvenience allows us on the other hand to
cache the calculation easily.