Delete replication entities
Recorder does not keep copies of musicbrainz entities anymore. That also
means, that the playback entity is *not* linked to any other first-party
entities anymore.
Currently, displaying a playback will be done by retrieving the view
data for each playback from mbs. This works with appropriate speed and
always shows up-to-date data from mbs.
In the future, we will obviously cache the information locally, so that
subsequent requests are much faster.