Add endpoint for batch import
This patch introduces a new endpoint playbacks/batch, on which the
client may POST a batch of resources rather than a single one. The
resource currently contains a single field "playbacks" which is a list
of regular playback resources.

The endpoint behaviour can be broken down into these simple facts:
1. If the API returns a non-200 HTTP code, *none* of the playbacks was
   stored.
2. In case of a 200 HTTP code, the response contains a field "results",
   which is a list of result items. The order of the POSTed list of
   playbacks defines the order of the result items. E.g. if the first
   playback was successfully persisted or not can be seen in the first
   result item. Each result item contains a boolean field "success".
3. 200 exit code is the default, even if every single operation in the
   batch failed. Only the success flag is meaningful. Non-200 exit codes
   happen only in non-expected error cases.
4. The order of persistence (if that even matters for the client) is the
   same as the order in the POSTed list.

The batch endpoint is meant to increase performance when importing large
amounts of data. This is why we do *not* provide the persisted resources
in the response, as it costs extra time to retrieve the view data.