Add --debug-dump-config flag
This flag makes cfm dump the entire config on the log output after it
has been loaded and all the command line flags have been added to the
runtime config. This can be useful to debug cfm. We do not print the
config by default, because it necessarily contains user credentials and
the risk is too high that someone accidentally pastes the log including
credentials.