Register bean to reset log level
This patch registers a bean "loglevel" during the setup of the logger.
The bean provides a sub that can be called to reset the log level.
Elsewhere in the code one can inject the bean if there's need to reset
the log level, which defaults to WARN.

In case that the log level was set via environment variable, we register
a NOP-sub that does nothing instead. That way we prevent that some code
resets the log level again, if the user specifically requested a
particular level.

We specifically do not provide the logger directly, because we want to
abstract the exact implementation of the logger.