Make Mpris2 connector a singleton and use it
This patch does several things:
1. it makes Cfm::Connector::Mpris2 a singleton so that we can inject it
   from the Cli
2. it changes Mpris2's listen method to accept a player name: call
   argument is better than state (imagine we reuse the singleton
   instance)
3. it injects Mpris2 in Cli and uses it there in an newly introduced
   command "record-mpris". For this command, there are also two new
   command line options.
   --player=something is mandatory and specifies the player's mpris2 bus
   name.
   --quiet or -q is optional and prevents resetting the log-level to
   "info" for recording mode. This option can obviously used in other
   commands as well and interpreted appropriately. The option itself
   does not trigger any "global" behaviour.

The Mpris2 connector is left mostly unchanged. It might be subject to
further improvements later. We should also consider to use a fail-safe
call to persist the playbacks, such that cfm does not fail if an error
occurs but rather keeps a local copy of the playback in a local file or
store until it can be retried.