Fix sorting bug
Due to the way we used to split up broken and non-broken playbacks, the
order was always destroyed by makePlaybackView(). This patch fixes the
code to maintain the order of all playbacks in the given list.

The patch also secretly includes a bugfix: for users without any
playbacks `cfm list` would cause a back-end error. The recorder service
would query mbs with an empty UUID list for producing a playback view.
Mbs would return 404 and the recorder code would fail. This is fixed by
simply not doing anything in case the playback list is empty.