Fix bad request bug for broken playbacks
In the change of the way mbs view data lookup works, I smuggled in a
bug. In order to fetch view data for one or multiple playbacks, we
request all recording (or release group) data for a list of recording
(or release group) UUIDs. If we have *only* broken playbacks to show,
this UUID list is empty. This happens whenever we just created a broken
playback or when we request the list of all broken playbacks.

Whenever the request list was empty, the REST call to mbs failed and
thus the request to the recorder also failed. I just added some fault
tolerance there: if the mbs request fails, we fallback to show original
view data. This is correct behaviour for broken playbacks and acceptable
behaviour for the case when mbs is not available.