Make threshold configurable
This patch gets the threshold (percentage of time to listen to a song in
order to consider it a playback) from the runtime config, instead of
hard-coding it into the player state machine. Firstly, we allow the user
to choose a threshold; secondly, this is very useful for debugging.

In order to retrieve the value from the config module, we inject the
config singleton into the PSM. This is the start of a patch series that
will refactor the PSM to be more readable and fit better into the
current architecture.