Add simple stop watch for state keeping
This patch adds a simple stop watch that supports only three actions:
start, stop, reset. The span from a start to a stop is called a trip.
The watch adds up all trips to a total duration, which is returned when
"reset" is called. Calling start or stop multiple times does not change
the state and thus not affect the result at all.

The stop watch will be used to track how long a song has been playing.
Whenever a state change of the player is noticed, the watch state can be
updated and - after a song change - the watch can be reset to obtain the
play time.