Add clean PlayerStateMachine
In order to replace the "old" PlayerStateMachine that firstly was
horrible to read and secondly missing some corner case, this patch
introduces a new cleaned up PlayerStateMachine. It utilises the
StopWatch and TrackState state keepers introduced in previous commits.
Instead of handling callbacks via sub refs, the new PlayerStateMachine
autowires a SendingEventHandler by default. If this behaviour is not
wanted, the client code can override this autowiring with any module of
the role PlayerEventHandler.

The commit adds a new option: gap-to-complete. The value is interpreted
as amount of seconds that may be missing in order to consider a track to
be completed. The reason behind this option is a corner case for singles
(only one track in a release) that spotify is too stupid to handle
explicitly with mpris2.
The option is not meant to be set by the user on a regular basis, which
is why it is not made available as command line flag.