Split connector and psm/eh into separate processes
Events from players are processed as follows:

+-----------+    +---------+    +----------+
| Connector + -> |   PSM   + -> |    EH    |
+-----------+    +---------+    +----------+

The connector gets events from a player in some way. The events contain
a player state (playing, paused, stopped) and track information. The
connector feeds those events into the player state machine (PSM) which
normalises the events (ignore redundant events etc) and derives
meaningful events for cfm: track started, track paused, track resumed,
track ended (completed) or track cancelled.

These high-level events are processed by calls to an event handler (EH).
In our case, handling the "ended event", for example (and in the future
also start and maybe cancel) includes REST calls against the back-end,
which may take a few hundred milliseconds.

Two facts:
1. this all used to happen in one process
2. the events from a player does not contain time stamps

lead to a discrepancy in time, One event might lead to a REST call in
the event handler. Another event quickly following the first one might
be processed much later and thus the timestamp that the connector puts
on this event is simply wrong.

To mitigate this effect, we run the connector in one process and the PSM
+ EH in another one. They are connected via a queue. The connector
includes a timestamp for all events. So even if the PSM/EH lag behind
several seconds, they process the events with a correct timestamp.

Fixes #31.