Add CSV and JSON formatters
Users might want to parse the output of cfm and ascii tables are hardly
appropriate as a parsable format. This patch adds two formatters, namely
CSV and JSON. The user can select them by specifying "csv" or "json" to
the --format flag.

In the code, we can just continue to singleton-inject the Formatter
module. The Formatter role is enhanced to figure out which formatter is
the right one (by querying the config) and the appropriate singleton
will be picked.

Fixes #38