Add json column for artists to RawPlaybackData
RawPlaybackData has an element collection for the artists. Due to
Hibernate being a serious clusterfuck, it is pretty hard to aggregate
playbacks over an element collection, or merely selecting it for that
matter. It mostly generates illegal SQL queries instead of complaining.

To simplify things, we add another redundant column "artist_json", which
simply contains a json encoded string list with all artists. Whether we
get rid of the element collection later is yet to be decideed. For the
time being, one can simply decide which one to use for queries, at the
acceptable cost of maintaining the redundant column.