Show accumulated broken playbacks
This patch enables the client to display data from the .../playbacks/acc
endpoint. This endpoint provides broken playbacks accumulated by their
values of artists/title/release. The result set is sorted by occurences,
i.e. if a single (broken) track has been played often, it will appear on
the top of this list. The returned accumulation is a good indicator as
to what playbacks should be fixed to reduce the number of broken ones
the most.
The data can be obtained by using the list command with the --acc flag.
All supported formatters are able to display the result set, with the
limitation that CSV only shows the first artist.