Make kv-store's input param configurable
This commit adds an optional parameter to the config module's kv_store()
sub. If not specified, it defaults to "set", which used to be the
hard-coded value for it. kv_store() produces a hash using multiple
arguments passed to set, which must have the pattern "key:value". The
user can specify -s multiple times and that way fill the kv-store.

This patch merely makes it possible to use another command line option
as source (other than --set/-s). We can use this to build kv-stores from
any (array) command line option.

Furthermore, the actual splitting logic is pulled out into its own sub,
as we will reuse this logic without the regular kv_store() soon.