Add magic -o option to set any config value
In regular use, we control what the user can write into the runtime
config by crafting command line parameters that only set exactly one
value. However, sometimes it can be very useful to allow setting of
arbitrary options, that are not in a predefined list.

This can be done by using the format -o option:value, so for example
instead of providing --import-source=lastfm it would be equivalent to
specify -o import-source:lastfm

Since the config module does not check if the *combination* of options
is valid, adding the magic "-o" does not cause any additional risk. In
the end, it is always the executing piece of code that decides which
parameters are relevant for it. Nonetheless, the option should be used
with care, as it also allows to override any hidden or otherwise set
options.

A further caveat must be mentioned: options set via -o are always
strings internally. Currently, I do not care about this caveat. However,
it must be noted, that this implies something like --broken is not
equivalent to -o broken:1
This is particularly important for options evaluated in a boolean
context. So it is not possible to set the option "broken" to false by
specifying -o broken:0, as this results in the string "0".