Introduce Fingerprints
This commit introduces the concept of Fingerprints in the backend. They
are a per-user mapping of
(artists, title, release) -> (release group, recording). Whenever a user
fixes a broken playback, the backend can record such a fingerprint in
order to detect the right release and recording automatically next time.

In future work we can make one user's fingerprints available to others
so that the community benefits from everyone's manual fixing.