Introducing wizard to fix playbacks interactively
This patch adds a first draft of a wizard that can be used to fix broken
playbacks and associate them with their musicbrainz identity. This is
not meant to be a final implementation but merely something that we can
work with to test the feature in general. Consequently, the usability
might not be all that great for now.