Allow update of accumulatd playbacks without actual fix
This commit changes the behaviour of POSTing on .../playbacks/acc.
Previously, it was not allowed to POST without setting the recordingId
and releasGroupId. We now allow this, in order to update the newly
introduced fixAttempt field. That way, clients can specify that an
attempt to fix an accumulation of playbacks has failed, probably because
there is no respective musicbrainz identiy for those playbacks. By
setting the fixAttempt field, we can identify playbacks that were
(attempted to be) fixed earlier and sort them to the end of the list.