Fix clear() in TrackState
It is kind of embarassing if you do the exact opposite of what you are
supposed to do. When clearing TrackState, it is as if it was moved into
pre-initialised state. So we should obviosly set pre_state to true and
not to false.

Aside from being just wrong, this also caused issues with spotify,
whenever one listened to a "single track release", because in this very
corner case, the PlayerStateMachine would actually need to clear
TrackState at the end of the track.