Wait for DBus service to show up
With this patch, the mpris2 connector *tries* to obtain the player's
DBus service instead of assuming it is just there. If the service is not
there, the connector waits for 'mpris-wait' seconds until it tries
again. This allows us to start cfm in record mode at any time, even if
the player is not yet running.