Skip Pause events in PSM if last_track is not yet set
This helps to ignore bogus signals from Spotify, which can produce an
invalid state. Spotify occasionally sends loads of Pause events, for
example right after startup. If cfm is running already, it will update
TrackState with the metadata from the pause events. However, since
*only* pause events occurred so far, PSM has never set track_state (why
should it??). If we now start playing anything but the track spotify
spammed around on startup, TrackState will report a change, but the PSM
claims there is no previous track. Consequently, some methods would fail
because they assumed that last_track is defined.

Ignoring the event like this should not hurt at all, but to be safe we
report the fact that we skip it to the user, so it’s easier to debug it.
Eventually, this reporting can be disabled maybe.

Fixes #35